Add a Windows user target
The idea is to build this with winelib and run it inside an ARM or ARM64
(or in theory any other) Wine build. Building this for native Windows
and e.g. run it on Windows RT would be nice, but I only have an ARM64
box and there is no public compiler yet that can create ARM64 windows
binaries. (They exist in the PSDK and Wine can run them though)

Wrappers for Windows DLLs will be maintained in an external repository
for starters.

This thing will need plenty of work to make signal handling working with
Windows APIs, handle process and thread creation the Windows way, etc.
Right now this compiles, but it will probably do something horribly
wrong as soon as an exception happens.