Rewrite path searching in the loader
The motivation for this was that we weren't skipping over dll's with the wrong
architecture like Wine does. In particular the common trick of placing both the
mingw32 and mingw64 sysroot's in the PATH doesn't work as expected and blows up
with weird hard-to-debug errors when using the one that's second in the PATH.
Fixing this implied refactoring the code substantially, because we now have to
implement searching ourselves and actually open the file and read the PE header
to check if it exists instead of relying on ... to do that for us. While in the
area, I took the liberty to refactor this to look more like wine's
implementation in ntdll/loader.c. Most of the new code was copied directly from
wine and works similarly, with the exception of read_nt_header() and the bit to
override the path for system32, which is now unified between the relative-path
and absolute-path cases.

One annoying thing is that we now read the header twice. We could probably pass
it around instead, similar to what wine does, but that would probably make
things more complicated.