Fix cmdline rewriting
The value returned by GetCommandLineW() is a cached version of what's in
the PEB, so when we allocate a new buffer and change the PEB pointer we
won't actually change what GetCommandLineW() returns, and when msvcrt
loads it then uses the value of GetCommandLineW() to initialize the
command line args. Do a similar hack to what we do with
GetCommandLineA() to make GetCommandLineW() return the right thing.